+++ .eslintrc.js
@@ -7,5 +7,5 @@ module.exports = {
     'plugin:@typescript-eslint/eslint-recommended',
     'plugin:@typescript-eslint/recommended',
     'prettier/@typescript-eslint',
-  ],
+  ]
 };
