+++ package.json
@@ -12,13 +12,13 @@
   "source": "src/signature_pad.ts",
   "main": "dist/signature_pad.umd.js",
   "module": "dist/signature_pad.js",
-  "types": "dist/types/signature_pad.d.ts",
+  "types": "dist/types/index.d.ts",
   "scripts": {
-    "build": "yarn run lint && yarn run clean && rollup --config && yarn run emit-types && yarn run update-docs",
+    "build": "npm run clean && rollup -c rollup.config.js",
+    "dev": "rollup -c rollup.config.js -w",
     "clean": "yarn run del dist",
-    "emit-types": "mkdir -p dist/types && yarn run tsc src/signature_pad.ts --lib DOM,ES2015 --declaration --declarationDir dist/types --emitDeclarationOnly",
     "format": "prettier --write '{src,tests}/**/*.{js,ts}'",
-    "lint": "eslint '{src,tests}/**/*.ts'",
+    "lint": "eslint src/**/*.ts tests/**/*.ts",
     "prepublishOnly": "yarn run build",
     "serve": "serve -l 9000 docs",
     "start": "yarn run build && yarn run serve",
@@ -48,9 +48,11 @@
     "husky": "^4.2.3",
     "jest": "^25.1.0",
     "lint-staged": "^10.0.8",
+    "miniprogram-api-typings": "^2.11.0-1",
     "prettier": "^2.0.1",
     "rollup": "^2.1.0",
     "rollup-plugin-terser": "^5.3.0",
+    "rollup-plugin-typescript2": "^0.27.1",
     "serve": "^11.3.0",
     "ts-jest": "^25.2.1",
     "tslib": "^1.11.1",
