+++ src/index.ts
@@ -1,0 +1,4 @@
+import SignaturePad from './signature_pad';
+import SignaturePadWechat from './signature_pad_wxa';
+
+export { SignaturePad, SignaturePadWechat };
