+++ src/signature_pad.ts
@@ -11,101 +11,24 @@
 
 import { Bezier } from './bezier';
 import { BasicPoint, Point } from './point';
-import { throttle } from './throttle';
+import SignaturePadBase, { Options, PointGroup } from './signature_pad_base';
 
-declare global {
-  // tslint:disable-next-line:interface-name
-  interface Window {
-    PointerEvent: typeof PointerEvent;
-  }
-}
-
-export interface Options {
-  dotSize?: number | (() => number);
-  minWidth?: number;
-  maxWidth?: number;
-  minDistance?: number;
-  backgroundColor?: string;
-  penColor?: string;
-  throttle?: number;
-  velocityFilterWeight?: number;
-  onBegin?: (event: MouseEvent | Touch) => void;
-  onEnd?: (event: MouseEvent | Touch) => void;
-}
-
-export interface PointGroup {
-  color: string;
-  points: BasicPoint[];
-}
-
-export default class SignaturePad {
-  // Public stuff
-  public dotSize: number | (() => number);
-  public minWidth: number;
-  public maxWidth: number;
-  public minDistance: number;
-  public backgroundColor: string;
-  public penColor: string;
-  public throttle: number;
-  public velocityFilterWeight: number;
+export default class SignaturePad extends SignaturePadBase {
   public onBegin?: (event: MouseEvent | Touch) => void;
   public onEnd?: (event: MouseEvent | Touch) => void;
+  public canvas: HTMLCanvasElement;
 
   // Private stuff
   /* tslint:disable: variable-name */
-  private _ctx: CanvasRenderingContext2D;
   private _mouseButtonDown: boolean;
-  private _isEmpty: boolean;
-  private _lastPoints: Point[]; // Stores up to 4 most recent points; used to generate a new curve
-  private _data: PointGroup[]; // Stores all points in groups (one group per line or dot)
-  private _lastVelocity: number;
-  private _lastWidth: number;
-  private _strokeMoveUpdate: (event: MouseEvent | Touch) => void;
-  /* tslint:enable: variable-name */
 
-  constructor(
-    private canvas: HTMLCanvasElement,
-    private options: Options = {},
-  ) {
-    this.velocityFilterWeight = options.velocityFilterWeight || 0.7;
-    this.minWidth = options.minWidth || 0.5;
-    this.maxWidth = options.maxWidth || 2.5;
-    this.throttle = ('throttle' in options ? options.throttle : 16) as number; // in milisecondss
-    this.minDistance = ('minDistance' in options
-      ? options.minDistance
-      : 5) as number; // in pixels
-    this.dotSize =
-      options.dotSize ||
-      function dotSize(this: SignaturePad): number {
-        return (this.minWidth + this.maxWidth) / 2;
-      };
-    this.penColor = options.penColor || 'black';
-    this.backgroundColor = options.backgroundColor || 'rgba(0,0,0,0)';
+  constructor(canvas: HTMLCanvasElement, options: Options = {}) {
+    super(canvas, options);
     this.onBegin = options.onBegin;
     this.onEnd = options.onEnd;
 
-    this._strokeMoveUpdate = this.throttle
-      ? throttle(SignaturePad.prototype._strokeUpdate, this.throttle)
-      : SignaturePad.prototype._strokeUpdate;
-    this._ctx = canvas.getContext('2d') as CanvasRenderingContext2D;
-
-    this.clear();
-
     // Enable mouse and touch event handlers
     this.on();
-  }
-
-  public clear(): void {
-    const { _ctx: ctx, canvas } = this;
-
-    // Clear canvas using background color
-    ctx.fillStyle = this.backgroundColor;
-    ctx.clearRect(0, 0, canvas.width, canvas.height);
-    ctx.fillRect(0, 0, canvas.width, canvas.height);
-
-    this._data = [];
-    this._reset();
-    this._isEmpty = true;
   }
 
   public fromDataURL(
@@ -118,10 +41,10 @@ export default class SignaturePad {
     const width = options.width || this.canvas.width / ratio;
     const height = options.height || this.canvas.height / ratio;
 
-    this._reset();
+    this.reset();
 
     image.onload = (): void => {
-      this._ctx.drawImage(image, 0, 0, width, height);
+      this.ctx.drawImage(image, 0, 0, width, height);
       if (callback) {
         callback();
       }
@@ -192,31 +115,31 @@ export default class SignaturePad {
       ({ color, point }) => this._drawDot({ color, point }),
     );
 
-    this._data = pointGroups;
+    this.data = pointGroups;
   }
 
   public toData(): PointGroup[] {
-    return this._data;
+    return this.data;
   }
 
   // Event handlers
   private _handleMouseDown = (event: MouseEvent): void => {
     if (event.which === 1) {
       this._mouseButtonDown = true;
-      this._strokeBegin(event);
+      this.strokeBegin(event);
     }
   };
 
   private _handleMouseMove = (event: MouseEvent): void => {
     if (this._mouseButtonDown) {
-      this._strokeMoveUpdate(event);
+      this.strokeMoveUpdate(event);
     }
   };
 
   private _handleMouseUp = (event: MouseEvent): void => {
     if (event.which === 1 && this._mouseButtonDown) {
       this._mouseButtonDown = false;
-      this._strokeEnd(event);
+      this.strokeEnd(event);
     }
   };
 
@@ -226,7 +149,7 @@ export default class SignaturePad {
 
     if (event.targetTouches.length === 1) {
       const touch = event.changedTouches[0];
-      this._strokeBegin(touch);
+      this.strokeBegin(touch);
     }
   };
 
@@ -235,7 +158,7 @@ export default class SignaturePad {
     event.preventDefault();
 
     const touch = event.targetTouches[0];
-    this._strokeMoveUpdate(touch);
+    this.strokeMoveUpdate(touch);
   };
 
   private _handleTouchEnd = (event: TouchEvent): void => {
@@ -244,72 +167,9 @@ export default class SignaturePad {
       event.preventDefault();
 
       const touch = event.changedTouches[0];
-      this._strokeEnd(touch);
+      this.strokeEnd(touch);
     }
   };
-
-  // Private methods
-  private _strokeBegin(event: MouseEvent | Touch): void {
-    const newPointGroup = {
-      color: this.penColor,
-      points: [],
-    };
-
-    if (typeof this.onBegin === 'function') {
-      this.onBegin(event);
-    }
-
-    this._data.push(newPointGroup);
-    this._reset();
-    this._strokeUpdate(event);
-  }
-
-  private _strokeUpdate(event: MouseEvent | Touch): void {
-    if (this._data.length === 0) {
-      // This can happen if clear() was called while a signature is still in progress,
-      // or if there is a race condition between start/update events.
-      this._strokeBegin(event)
-      return
-    }
-
-    const x = event.clientX;
-    const y = event.clientY;
-
-    const point = this._createPoint(x, y);
-    const lastPointGroup = this._data[this._data.length - 1];
-    const lastPoints = lastPointGroup.points;
-    const lastPoint =
-      lastPoints.length > 0 && lastPoints[lastPoints.length - 1];
-    const isLastPointTooClose = lastPoint
-      ? point.distanceTo(lastPoint) <= this.minDistance
-      : false;
-    const color = lastPointGroup.color;
-
-    // Skip this point if it's too close to the previous one
-    if (!lastPoint || !(lastPoint && isLastPointTooClose)) {
-      const curve = this._addPoint(point);
-
-      if (!lastPoint) {
-        this._drawDot({ color, point });
-      } else if (curve) {
-        this._drawCurve({ color, curve });
-      }
-
-      lastPoints.push({
-        time: point.time,
-        x: point.x,
-        y: point.y,
-      });
-    }
-  }
-
-  private _strokeEnd(event: MouseEvent | Touch): void {
-    this._strokeUpdate(event);
-
-    if (typeof this.onEnd === 'function') {
-      this.onEnd(event);
-    }
-  }
 
   private _handlePointerEvents(): void {
     this._mouseButtonDown = false;
@@ -333,137 +193,6 @@ export default class SignaturePad {
     this.canvas.addEventListener('touchend', this._handleTouchEnd);
   }
 
-  // Called when a new line is started
-  private _reset(): void {
-    this._lastPoints = [];
-    this._lastVelocity = 0;
-    this._lastWidth = (this.minWidth + this.maxWidth) / 2;
-    this._ctx.fillStyle = this.penColor;
-  }
-
-  private _createPoint(x: number, y: number): Point {
-    const rect = this.canvas.getBoundingClientRect();
-
-    return new Point(x - rect.left, y - rect.top, new Date().getTime());
-  }
-
-  // Add point to _lastPoints array and generate a new curve if there are enough points (i.e. 3)
-  private _addPoint(point: Point): Bezier | null {
-    const { _lastPoints } = this;
-
-    _lastPoints.push(point);
-
-    if (_lastPoints.length > 2) {
-      // To reduce the initial lag make it work with 3 points
-      // by copying the first point to the beginning.
-      if (_lastPoints.length === 3) {
-        _lastPoints.unshift(_lastPoints[0]);
-      }
-
-      // _points array will always have 4 points here.
-      const widths = this._calculateCurveWidths(_lastPoints[1], _lastPoints[2]);
-      const curve = Bezier.fromPoints(_lastPoints, widths);
-
-      // Remove the first element from the list, so that there are no more than 4 points at any time.
-      _lastPoints.shift();
-
-      return curve;
-    }
-
-    return null;
-  }
-
-  private _calculateCurveWidths(
-    startPoint: Point,
-    endPoint: Point,
-  ): { start: number; end: number } {
-    const velocity =
-      this.velocityFilterWeight * endPoint.velocityFrom(startPoint) +
-      (1 - this.velocityFilterWeight) * this._lastVelocity;
-
-    const newWidth = this._strokeWidth(velocity);
-
-    const widths = {
-      end: newWidth,
-      start: this._lastWidth,
-    };
-
-    this._lastVelocity = velocity;
-    this._lastWidth = newWidth;
-
-    return widths;
-  }
-
-  private _strokeWidth(velocity: number): number {
-    return Math.max(this.maxWidth / (velocity + 1), this.minWidth);
-  }
-
-  private _drawCurveSegment(x: number, y: number, width: number): void {
-    const ctx = this._ctx;
-
-    ctx.moveTo(x, y);
-    ctx.arc(x, y, width, 0, 2 * Math.PI, false);
-    this._isEmpty = false;
-  }
-
-  private _drawCurve({ color, curve }: { color: string; curve: Bezier }): void {
-    const ctx = this._ctx;
-    const widthDelta = curve.endWidth - curve.startWidth;
-    // '2' is just an arbitrary number here. If only lenght is used, then
-    // there are gaps between curve segments :/
-    const drawSteps = Math.floor(curve.length()) * 2;
-
-    ctx.beginPath();
-    ctx.fillStyle = color;
-
-    for (let i = 0; i < drawSteps; i += 1) {
-      // Calculate the Bezier (x, y) coordinate for this step.
-      const t = i / drawSteps;
-      const tt = t * t;
-      const ttt = tt * t;
-      const u = 1 - t;
-      const uu = u * u;
-      const uuu = uu * u;
-
-      let x = uuu * curve.startPoint.x;
-      x += 3 * uu * t * curve.control1.x;
-      x += 3 * u * tt * curve.control2.x;
-      x += ttt * curve.endPoint.x;
-
-      let y = uuu * curve.startPoint.y;
-      y += 3 * uu * t * curve.control1.y;
-      y += 3 * u * tt * curve.control2.y;
-      y += ttt * curve.endPoint.y;
-
-      const width = Math.min(
-        curve.startWidth + ttt * widthDelta,
-        this.maxWidth,
-      );
-      this._drawCurveSegment(x, y, width);
-    }
-
-    ctx.closePath();
-    ctx.fill();
-  }
-
-  private _drawDot({
-    color,
-    point,
-  }: {
-    color: string;
-    point: BasicPoint;
-  }): void {
-    const ctx = this._ctx;
-    const width =
-      typeof this.dotSize === 'function' ? this.dotSize() : this.dotSize;
-
-    ctx.beginPath();
-    this._drawCurveSegment(point.x, point.y, width);
-    ctx.closePath();
-    ctx.fillStyle = color;
-    ctx.fill();
-  }
-
   private _fromData(
     pointGroups: PointGroup[],
     drawCurve: SignaturePad['_drawCurve'],
@@ -482,7 +211,7 @@ export default class SignaturePad {
           this.penColor = color;
 
           if (j === 0) {
-            this._reset();
+            this.reset();
           }
 
           const curve = this._addPoint(point);
@@ -492,7 +221,7 @@ export default class SignaturePad {
           }
         }
       } else {
-        this._reset();
+        this.reset();
 
         drawDot({
           color,
@@ -503,7 +232,7 @@ export default class SignaturePad {
   }
 
   private _toSVG(): string {
-    const pointGroups = this._data;
+    const pointGroups = this.data;
     const ratio = Math.max(window.devicePixelRatio || 1, 1);
     const minX = 0;
     const minY = 0;
